+++ src/masternode-payments.h
@@ -8,8 +8,8 @@
 #include "util.h"
 #include "core_io.h"
 #include "key.h"
-#include "main.h"
 #include "masternode.h"
+#include "net_processing.h"
 #include "utilstrencodings.h"
 
 class CMasternodePayments;
