+++ src/test/main_tests.cpp
@@ -5,6 +5,7 @@
 
 #include "chainparams.h"
 #include "main.h"
+#include "net.h"
 
 #include "test/test_dash.h"
